hello_ext: use size_t for input data size instead of int

+++ lib/hello_ext.h
@@ -33,7 +33,7 @@
 int _gnutls_parse_hello_extensions(gnutls_session_t session,
 				   gnutls_ext_flags_t msg,
 				   gnutls_ext_parse_type_t parse_type,
-				   const uint8_t *data, int data_size);
+				   const uint8_t *data, size_t data_size);
 int _gnutls_gen_hello_extensions(gnutls_session_t session,
 				 gnutls_buffer_st *extdata,
 				 gnutls_ext_flags_t msg,
